Work around Carthage braindeath (#838)
Carthare has this weird ardency to build every Xcode project in can find
in the source tree instead of working only with the one in the root dir.

There is probably some deep, well-intentioned reason behind that, but it
also causes issues for us. Carthage will end up visiting all the example
projects, some of which use SPM to pull Themis dependency, and each of
them will do so independently, causing builds to time out.

We are miraculously saved by Themis being distributed as binaries, so
Carthage and SPM at least fetch the binary and don't go deeper. If we
have supported only from-source builds, this would have led to infinite
recursion.

There is not much that we can do about it, but at least we can keep CI
builds in shape with a workaround. If all distractions are removed,
Carthage will focus on the Xcode project it is supposed to work with.
This at least lets out CI to function.

Users that instruct Carthage to build dependencies from source
(via --no-use-binaries) will still experience this issue of extremely
long build times and occasional failures. If that's the case for you,
please accept my condolences, go complain in Carthage issue tracker,
and maybe think about choosing a different line of work.