Use Ninja for BoringSSL if available (#837)
* Detect Ninja in configure

Look into the system if there is "ninja" binary available in PATH.
If there is one, note it for future use.

* Use Ninja for BoringSSL if available

Ninja build tool significantly reduces compilation time for BoringSSL in
default configuration. This is mostly due to the fact that it performs
parallel builds by default but there are other gains (Makefiles that
CMake generates are not that optimal). "make -j4" is also a viable
option, but it often ends up brittle.

* Install ninja for jobs that build BoringSSL

Make sure it's actually installed where relevant.

* Make a quick note in changelog about Ninja support