Prefer GitHub Discussions for general inquiries
This feature is still technically in beta but it's already usable.
That's a nice way to keep general "question" issues out of the issue
tracker which should contain only something actionable.

Make Discussions the first link that you'd like to click if you just
want to chat about something and not to report a bug.