Reference implementation of Themis
After a while I have realized that putting reference implementation into
the same repository as documentation is probably not the greatest idea.
Let's move it into the warm embrace of the monorepo, where we can run
tests on it and evolve the code together with the "real" GoThemis.

Start by setting a new top-level directory so that Go imports look nice:

    github.com/cossacklabs/themis/reference/...

That's where reference implementation will live.