Run "make test_go" in non-modular, GOPATH mode
Currently "make test_go" is similar to script languages in that it tests
global installation of GoThemis. However, Go has been steadily migrating
away from global package installation since Go 1.11 and the advent of
modules. Starting with Go 1.16 the modular approach is the default.

Keep the existing behavior of "make test_go" which expects global
installation of GoThemis. It will not work without GoThemis installed
into GOPATH via

    GO111MODULE=off go get github.com/cossacklabs/themis/gothemis

I don't know whether it's time to deprecate this Makefile target or not.
It still kinda works, but it's definitely less useful than it was.