Modular ES6-style "tools" for WasmThemis
In addition to ES5-style tools add those that use ES6 module syntax.
This is literally the same code, but with different import syntax.
Modern Node.js can work with both. (And make sure it does on CI.)

If we are going that route of compiling from TypeScript, these will be
the only tools that are going to ultimately work.