Code examples for WasmThemis (ES5-style for Node.js)
WasmThemis did not really have any code examples. We bullshit our way in
by pretending that JsThemis examples are good, but they actually are not
compatible with WasmThemis due to API differences and WebAssembly stuff.

Write examples for WasmThemis, start with ES5-style for Node.js. This is
basically copypasta of JsThemis examples, just cleaned up a bit. I don't
want to invent CI tests for new examples so they are compatible.