Fix variable references in Secure Cell + passphrase tool
This fails in ES6 mode but not in ES5, that's because ES6 is a bit more
strict on what gets captured and survives and what does not. Note that
assigning previously undeclared variable *is not* an error in JavaScript
(this is JavaScript, after all), but it *depends* whether that variable
will be captured (this is JavaScript, you know).