Example WasmThemis project with webpack
Finally my dream is coming true: interactive example of Themis right in
the browser. It's a bit sad that I have to work for my own dreams, but
doing what I can with what I got.

The first commit adds a skeleton with webpack 5, pulling in WasmThemis
and half of the internet of other dependencies. Don't think about it too
much, okay? This is normal for JavaScript.

Right now we don't do anything, but at least you can see that WasmThemis
has loaded successfully. Or not.