Apply Cascading Style Shits
These took a while to get them more or less right. I don't envy web
designers that have to deal with this shit on a daily basis. I am so
exhausted from Safari thinking differently™ that I am not going to
clean up the CSS file, just throwing it in as is. (Like, why the hell
disabled inputs are 1 pixel shorter that normal ones? Who knows.)

But at least now the page looks *somewhat* decently. I don't care for
more formatting, as long as it's functional.