WasmThemis code examples (#846)
* Test WasmThemis command-line tools

Put it under "Code examples" job because that's the only example code
for WasmThemis that we have at the moment.

* Modular ES6-style "tools" for WasmThemis

In addition to ES5-style tools add those that use ES6 module syntax.
This is literally the same code, but with different import syntax.
Modern Node.js can work with both. (And make sure it does on CI.)

If we are going that route of compiling from TypeScript, these will be
the only tools that are going to ultimately work.

* Code examples for WasmThemis (ES5-style for Node.js)

WasmThemis did not really have any code examples. We bullshit our way in
by pretending that JsThemis examples are good, but they actually are not
compatible with WasmThemis due to API differences and WebAssembly stuff.

Write examples for WasmThemis, start with ES5-style for Node.js. This is
basically copypasta of JsThemis examples, just cleaned up a bit. I don't
want to invent CI tests for new examples so they are compatible.

* Code examples for WasmThemis (ES6-style for Node.js)

Now the same thing but with different file name extension, different
import syntax, and async function support (though, Node.js APIs are
still mostly callback-based).

* Fix variable references in Secure Cell + passphrase tool

This fails in ES6 mode but not in ES5, that's because ES6 is a bit more
strict on what gets captured and survives and what does not. Note that
assigning previously undeclared variable *is not* an error in JavaScript
(this is JavaScript, after all), but it *depends* whether that variable
will be captured (this is JavaScript, you know).

* Build WasmThemis only once

I know that GitHub Actions compute is kinda free, but still... building
WasmThemis eight times for all our jobs is wasteful. We do this in
exactly the same Emscripten environment, host Node.js version does not
affect the build result.

Split out the building part into a separate job, run Themis Core tests
there too (with Emscripten's Node.js). Then archive the build directory
and upload it as an artifact to be shared with other jobs.

Other jobs that run WasmThemis unit-tests and examples will pull the
build directory with WebAssembly build results, unpack it, and run then
stuff with prebuilt WebAssembly binaries (use that portability for once).
This is similar to what users will do: download binaries from npm.

Note that unit tests and examples *are* exercised with various Node.js
versions.

* Make ES5-like code look like ES5

That is, no using "let" and "const" which are actually ES6-only and
are allowed in code executed by Node.js only because Node.js.

Am I doing this right, mom?

* Run WasmThemis jobs when examples change

* Fail tests on unhandled rejections

Some of our examples don't handle top-level promise rejection because
it's bothersome. Make sure that Node.js process exits with a failure
if those examples fail due to an unhandled promise rejection.

That is, code like this will now fail tests:

    themis.initialized.then(function() {
        throw new Error('this should cause non-zero exit code')
    })

instead of just printing a warning.

NODE_OPTIONS can include "--unhandled-rejections" only for Node.js v10
or later. v8 will fail due to an unknown option. And we still test v8.

GitHub Actions do not provide a nice way to set environment variables
based on the matrix configuration and I don't want to add a job step
that does that manually.

We care about this setting only when testing examples, so set it just
there, where we use Node.js v12+