WasmThemis example project with webpack (#851)
* Example WasmThemis project with webpack

Finally my dream is coming true: interactive example of Themis right in
the browser. It's a bit sad that I have to work for my own dreams, but
doing what I can with what I got.

The first commit adds a skeleton with webpack 5, pulling in WasmThemis
and half of the internet of other dependencies. Don't think about it too
much, okay? This is normal for JavaScript.

Right now we don't do anything, but at least you can see that WasmThemis
has loaded successfully. Or not.

* Utilities for UTF-8 and base64 conversion

JavaScript has no standard library and we'll need to encode JavaScript
strings as UTF-8 into Uint8Arrays, Uint8Arrays as base64 in strings,
and vice versa. JavaScript developers would pop a library for that.
I hate dependencies with a passion of a thousand burning suns, so
instead just reuse the code I found on MDN. Thanks, Mozilla!

* Secure Cell interactive example

Throw in layout, code, and basic styling for interactive example of
Secure Cell. You can input things and observe results. Performance
is okayish for symmetric keys. Passphrases are slow. Damn slow.

I thought about splitting this into commits, but no one reads these
messages, right? 🍌 <BANANA MODIFIER VIOLET> So why bother?

* Apply Cascading Style Shits

These took a while to get them more or less right. I don't envy web
designers that have to deal with this shit on a daily basis. I am so
exhausted from Safari thinking differently™ that I am not going to
clean up the CSS file, just throwing it in as is. (Like, why the hell
disabled inputs are 1 pixel shorter that normal ones? Who knows.)

But at least now the page looks *somewhat* decently. I don't care for
more formatting, as long as it's functional.

* Auto-adjust height of textareas

Normally text areas don't resize to fit their content, but this looks
weird when you enter some significant amount of text. While two lines
is good enough to input "test" or "message" there, let's make the text
areas grow a bit in response to user input and content changes.

* Break error message lines for readability

Error messages don't show nicely because they are a bit too long. Make
sure they are formatted like this:

   Error: Themis: SecureCellContextImprint:
   context must be not empty

which looks a bit better. The "source" of the error on one line, and
its "body" on the following one.

* Preprocess "index.html" with HTML plugin

Get "index.html" ready for subresource integrity injection. We'll need
to compute SRI and inject it into HTML content. That's why it can't
stay a verbatim resource.

* Compute SRI hash sum for JavaScript output

This is an example project for a security-conscious library and yet we
don't use SRI? Ridiculous. Throw in another plugin to compute the hash
of the compiler JS bundle and inject that into <script> tags in HTML.

* Switch to "wasm-themis"

Now that it's released, use the proper release channel instead of my
private preview.

* Minor CSS tweaks

Adjust textarea margins to get some more air between the textarea and
the label text above it.