Bring back "initialized" promise with some vile hacks
Historically, WasmThemis has exported only the "initialized" promise.
WasmThemis was not modularized and kicked off "libthemis.wasm" download
right after the module is loaded. Moreover, there was no way to use
a different path to WebAssembly code: it had to be located in a file
named "libthemis.wasm" next to the JavaScript code bundle.

TypeScript rewrite introduced an alternative interface "initialize()"
which allows to specify the URL. However, this required WasmThemis
modularization. That is, WasmThemis *will not* be initialized unless
initialize() is called.

This introduces two issues with backwards compatibility:

  1. "initialize" promise was gone.

  2. You can no longer expect that WasmThemis will be downloaded
     in background without awaiting any promise.

Due to modularization, there is no (practical) way to fix the second
issue. Just don't do, it's a bad practice.

However, absence of "initialized" is a real breaking change that will
break well-formed applications.

Reintroduce the "initialized" promise and make sure it works by calling
the "initialize()" function behind the scenes. It's a bit of a hack, but
it works, so whatever.

Note that "initialize" is *not* deprecated or anything. It's a perfectly
valid API if you don't need to specify a custom path to "libthemis.wasm"