Allow calling `initialize()` only once (#857)
WasmThemis initialization fills in the "context.libthemis" singleton
that is actually used by all WasmThemis functions. That is, calling
initialize() multiple times is technically permitted, but it might not
have an effect that you expect. WasmThemis will always use WebAssembly
context from the last initialization.

While reinitialization might be useful, it does not seem to be now.
Let's allow calling initialize() only once. (This includes calling it
indirectly via the "initialized" promise.) We can always allow it later,
once the semantics of this are clarified.