Provide default export from the main module
Because WasmThemis 0.13.X has been using CommonJS modules, various
transpilers interepreted

    import themis from 'wasm-themis';

as a valid way to import WasmThemis stuff under "themis" namespace
in genuine ES6 code.

The "correct" way to do this is as follows:

    import * as themis from 'wasm-themis';

but we need to make sure that the previous form still works.

It's awkward, but when did "it looks awful" was an arguments against
making a breaking change?