Keep a copy of libthemis.wasm in "src" (#859)
That's where this file is located in WasmThemis 0.13.X and some users
might have hardcoded its location (like it's hardcoded in WasmThemis
example with webpack, for example). Let's package the original copy
of libthemis.wasm at its original path, for compatibility.

This makes installed, unpackaged size a bit larger, but JavaScript
developers are used to "node_modules" being heavier then their mom
so it's okay. And this does not affect downloads, since identical
files are compressed with deduplication.