Provide default export from the main module (#860)
* Provide default export from the main module

Because WasmThemis 0.13.X has been using CommonJS modules, various
transpilers interpreted

    import themis from 'wasm-themis';

as a valid way to import WasmThemis stuff under "themis" namespace
in genuine ES6 code.

The "correct" way to do this is as follows:

    import * as themis from 'wasm-themis';

but we need to make sure that the previous form still works.

It's awkward, but when did "it looks awful" was an arguments against
making a breaking change?

* Drop unnecessary "secure_cell.ts"

"index.ts" exports all Secure Cell variations individually and this file
is not really used anywhere else. File structure of WasmThemis is
implementation detail. You're not supposed to peek into it.