rust: Avoid double references
Recent Clippy started complaining when you pass a reference to
reference.  Compiler will automatically dereference those, but
the warnings suggest it's redundant. Remove those. Originally,
I wrote it like that for consistency with other arguments, but
that would better be solved with a different argument parser.

Anyhow, stop offending our dear stationery and unbreak the build.