CI: Install OpenSSL 1.1 via Homebrew
Homebrew formula "openssl" now defaults to "openssl@3" which is not
currently supported by Themis (so much that the tests are broken).
While we certainly can compile Themis, it doesn't look like it's
working properly. First of all  start installing OpenSSL 1.1 on macOS
where it's already the default (with Homebrew, that is).