soter: Fail the build with OpenSSL 3.0
As I said in the comment, it's pretty risky at the moment to build
Themis against OpenSSL 3.0, so let's explicitly disallow it.

This applies to previously released versions too, but I can't be
bothered to issue a hotfix for that. Binary releases for Linux are not
affected (they still use OpenSSL 1.1), Homebrew release for macOS will
be patched up to use OpenSSL 1.1 separately, and whoever is brave enough
to build Themis from source manually -- you will be warned in docs.