core: Choose `openssl@1.1` from Homebrew by default (#871)
* CI: Install OpenSSL 1.1 via Homebrew

Homebrew formula "openssl" now defaults to "openssl@3" which is not
currently supported by Themis (so much that the tests are broken).
While we certainly can compile Themis, it doesn't look like it's
working properly. First of all  start installing OpenSSL 1.1 on macOS
where it's already the default (with Homebrew, that is).

* configure: Choose "openssl@1.1" from Homebrew by default

When automatically detecting OpenSSL location, choose OpenSSL 1.1.
This is applicable only for from-source builds. Technically, all
previously released source tarballs fail to build right now, but
that's not much of an issue since not many macOS users build Themis
from source. The currently published Homebrew formula is broken too.
That will be patched up separately.

* changelog: Note that we use OpenSSL 1.1 and not 3.0