core: Deny building with OpenSSL 3.0 (#872)
* soter: Fail the build with OpenSSL 3.0

As I said in the comment, it's pretty risky at the moment to build
Themis against OpenSSL 3.0, so let's explicitly disallow it.

This applies to previously released versions too, but I can't be
bothered to issue a hotfix for that. Binary releases for Linux are not
affected (they still use OpenSSL 1.1), Homebrew release for macOS will
be patched up to use OpenSSL 1.1 separately, and whoever is brave enough
to build Themis from source manually -- you will be warned in docs.

* CI: Verify that OpenSSL 3.0 build fails

Throw in an extra step to ensure that OpenSSL 3.0 is indeed failing the
build. At the moment only macOS Homebrew seems to have OpenSSL 3.0
available, so test with that. Linux distros are currently working on
transitions too, but testing just one OS should be fine at the moment.

* changelog: Mention compilation error in the changelog

The intention here is that if we release Themis 0.14 as is, without
OpenSSL 3.0 support, then people should be aware of this #error
directive that will fail their builds.

* Makefile: Allow to opt-in OpenSSL 3 usage for experts

Provide a not-really-documented way to build Themis with OpenSSL 3
by setting environment variable WITH_EXPERIMENTAL_OPENSSL_3_SUPPORT.