RSA: keygen: Set status code before goto
Another convention is that immediately before "goto" you set the status
code that the function returns after cleaning up, unless the previous
line has a function call that returns a status code that be bubble up.

Yes, the same status code is often already set at the beginning of the
function. That is completely different thing -- that's giving variables
initial values. Compiler is free to optimize those away if it knows --
and it almost always knows that the "res" value didn't change. Or maybe
it did, then in that case we will be returning the status code we want.