RSA: keygen: Deobfuscate status check
Previously this expression did not free the "ctx" is cleanup failed.
Of course, if you read the code above you'll see that it never fails
(considering the early check that we make). However, if it did, that
would be a leak, because callers of soter_rsa_key_pair_gen_destroy()
expect the key generation context to be dellocated regardless of the
returned status code.