RSA: keygen: Unify return path
The asymmetry here ticks me off a bit, so let's unify this piece,
following the tacit rule that functions should "return" either early
during parameter checks, or once at the end. Generally, this is the
safest thing to do in C, making sure that resource deallocation and
error handling are taken care of.