sign/verify: Introduce EVP_PKEY into soter_sign_ctx_t
Turns out, Soter generates EC keys incorrectly, by misusing OpenSSL API.
It somehow works out for OpenSSL 1.1, but in OpenSSL 3.0 this fails.

I suspect the issue stems from a misconception that EVP_PKEY_CTX is
a container for generated keypair as EVP_PKEY. It is not. While it's
true that EVP_PKEY_CTX_new() accepts an EVP_PKEY which can then be
retrived using EVP_PKEY_CTX_get0_pkey() -- this is not a keypair.
For EC keys this EVP_PKEY is "key parameters" (the curve, etc.)
Generated keypair is output into a *different* EVP_PKEY, and it
should be generated with a different function from that Soter uses.

The same is mostly true for RSA as well. While it *happens so* that
generated RSA key gets placed into EVP_PKEY_CTX, it should be a separate
EVP_PKEY object. It works out only because RSA does not have parameters.

Fixing this thing is going to be involved. The first bunch of patches
only divorces Soter from the notion that EVP_PKEY_CTX is a container
for keypairs and introduces EVP_PKEY fields for storing generated
keypairs, independently from EVP_PKEY_CTX used to generate them.

This patch in particular takes care of the sign/verify context.
First we just add the field, initialize it constructors, and free it
in destructors of soter_sign_ctx_t.