sign/rsa: Return generated EVP_PKEY via out-parameter
Instead of using EVP_PKEY_CTX as a container for EVP_PKEY to be filled
in during key generation, accept an out-parameter for EVP_PKEY,
following the signature of EVP_PKEY_keygen() which is actually used
for generating the key (and gets passed the out-parameter now).

Currently the NULL checks we perform on dereferenced EVP_PKEY do not
make much sense, but I'm retaining them until a later point where they
could be safely removed.