sign/rsa: Introduce "err" label and code path
Start building the error handling code path here by introducing an "err"
label to which we'll jump on error.

"pub_exp" needs a special treatment here since it's passed to the
EVP_PKEY_CTX. So if we fail to pass it then we free it, but otherwise
we must NOT free it again on future error handling. Thus reset it to
NULL (which is safe to pass to BN_free()).