sign/rsa: Move EVP_PKEY allocation to key generation
Since we generate a new key, might as well allocate it there too.

We immediately assign the allocated key to ppkey and the caller will
free that for us if we fail. That's also why we can also return before
we allocate EVP_PKEY_CTX and don't need to "goto err". (Though, both
these points are moot since these lines are going away soon.)