sign/rsa: Use EVP_PKEY_CTX_new_id() for key generation
If you read OpenSSL documentation *very carefully*, you'd see that
the correct way to generate keys is to start with EVP_PKEY_CTX allocated
by EVP_PKEY_CTX_new_id() with the key type you want to generate.
The sequence we do here is technically incorrect and just happens to
sorta-kinda work out for RSA because of the implementation quirks.

The other form -- EVP_PKEY_CTX_new() -- takes EVP_PKEY with
"key parameters", which there are none for RSA. This will be important
for EC key generation later, but not here. I'll rant about it when
I will be dealing with EC code paths.