sign/ec: Use EVP_PKEY_keygen() for key generation
And now start reading very carefully.

TTTTT H   H III  SSS    III  SSS    V   V III TTTTT   A   L
  T   H   H  I  S        I  S       V   V  I    T    A A  L
  T   HHHHH  I   SSS     I   SSS    V   V  I    T   AAAAA L
  T   H   H  I      S    I      S    V V   I    T   A   A L
  T   H   H III  SSS    III  SSS      V   III   T   A   A LLLLL

The following several commits are the reason for this entire
giant patchset in the first place.

This is how you generate EC keys with OpenSSL.

Not how they were generated before, but in this way.

Previous way sorta-kinda works with OpenSSL 1.1 but it got broken in
OpenSSL 3.0 and OpenSSL developers have confirmed that it should have
never worked in the first place, but it did, and people copied and
pasted incorrect snippets all around, and that's how this misconception
got born. Because of course OpenSSL 1.1 documentation provides zero
examples of how to generate EC keys. OpenSSL 3.0 has improved. A bit.

First of all, there is a dedicated function for generating the keypair.
The context also needs to be initialized for key generation first.
(Actually, you need *a different* EVP_PKEY_CTX for this, but that's
going to be handled in the following commits.)