sign/ec: Use EVP_PKEY_CTX_new_id() for parameter generation
Also, you should use EVP_PKEY_CTX_new_id() to allocate EVP_PKEY_CTX
for parameter generation. *Not* use an empty EVP_PKEY with only
the type set. This works in OpenSSL 1.1 by accident but this is not
how parameter generation is supposed to be performed. (And OpenSSL 3.0
breaks EVP_PKEY_CTX_new() usage for EVP_PKEY_paramgen() completely.)

With this change, sign/verify tests finally pass with OpenSSL 3.0,
which is the goal of the patchset.

The following commits are mopping up and filling in some blanks.