CI: Work around legacy Node.js 8.x issues
As described in referenced issue, Node.js 8.x tends to cause crashes
because it embeds OpenSSL 1.0.2 when Themis has been built against
OpenSSL 1.1.1. It seems before all these changes for OpenSSL 3.0
compatibility we somehow avoided calling any APIs that have actually
broken binary compatiblity between 1.0.2 and 1.1.1, but now we do.

The documented workaround is to use BoringSSL build so BoringSSL we use.
That's going to check this workaround in the first place, lol.