ka: Introduce EVP_PKEY into soter_asym_ka_t
And again, Soter makes the same mistake of using EVP_PKEY_CTX as
a container for EVP_PKEY with unspecified role and retaining this
context. In fact, soter_asym_ka_t deals with three separate operations
each of which requires different contexts:

  1. EC parameter generation (input: curve, output: parameters).
  2. EC keypair generation (input: parameters, output: keypair).
  3. Shared secret derivation (input: keypair, peer public key,
     output: shared secret).

And all of these need separate contexts, not just one being reused.
Previous incorrect code worked with OpenSSL 1.1.1 and before by
accident, but with OpenSSL 3.0 this fails. We need to do things
correctly now.

Start by introducing fields for the key parameters and the keypair.