Unify minor versions across wrappers (#878)
* wasm: Roll package version back to 0.13.XX

It got pushed over to 0.14 in anticipation of the next release which
never came on time. And now the internal CI tooling demands version
to be 0.13.XX across all wrappers because reasons. Do so then.
The previously released version is 0.13.1, use the next one.

* Themis.xcodeproj: Unify MARKETING_VERSION = 0.13.12

I have no clue how it turned out that some targets use 0.13.8 and some
use 0.13.12 but this confuses internal tooling that expects consistent
version across all targets. Use the latest one then.