rust: Add Send bound to SecureSessionTransport trait
Parent struct SecureSession expects to implement Send at all times.
However, it would be unsound if SecureSessionTransport does not
implement Send. This is a blunder in the original implementation.

Well, let's hope that all reasonable implementations are already Send,
and just require SecureSessionTransport to be Send. I'd rather make
0.15.1 later with some ugly generic fix if some users show up with
a genuine use-case for !Send transport callbacks.

That said, SecureSession could be made to implement Send based on the
type of SecureSessionTransport it uses. However, I don't want to get
generics into the mix unless I have to, since that might turn out
to be another breaking change.