rust: unsafe impl Send for SecureSessionContext
Because the first "unsafe impl" went so well, let's add another one.
This makes sure that the first "unsafe impl" is well-founded: that is,
all fields are actually Send -- compiler does not check that, trusting
our "unsafe" but recent Clippy does, and it complained about it.

With "Box<dyn SecureSessionTransport>" being Send now, the only
remaining thing that is !Send is secure_session_user_callbacks_t,
because of that "void* user_context" pointer in it. The pointer
points to SecureSessionContext struct itself, which make it not only
unsafe to send across threads, it's unsafe to move it at all. That's
why we put it in a Box which stays in the same place on heap until
SecureSession is dropped.

I believe this should be safe now.