rust: Add Send bound to SecureSessionTransport trait (#898)
* rust: Add Send bound to SecureSessionTransport trait

Parent struct SecureSession expects to implement Send at all times.
However, it would be unsound if SecureSessionTransport does not
implement Send. This is a blunder in the original implementation.

Well, let's hope that all reasonable implementations are already Send,
and just require SecureSessionTransport to be Send. I'd rather make
0.15.1 later with some ugly generic fix if some users show up with
a genuine use-case for !Send transport callbacks.

That said, SecureSession could be made to implement Send based on the
type of SecureSessionTransport it uses. However, I don't want to get
generics into the mix unless I have to, since that might turn out
to be another breaking change.

* rust: unsafe impl Send for SecureSessionContext

Because the first "unsafe impl" went so well, let's add another one.
This makes sure that the first "unsafe impl" is well-founded: that is,
all fields are actually Send -- compiler does not check that, trusting
our "unsafe" but recent Clippy does, and it complained about it.

With "Box<dyn SecureSessionTransport>" being Send now, the only
remaining thing that is !Send is secure_session_user_callbacks_t,
because of that "void* user_context" pointer in it. The pointer
points to SecureSessionContext struct itself, which make it not only
unsafe to send across threads, it's unsafe to move it at all. That's
why we put it in a Box which stays in the same place on heap until
SecureSession is dropped.

I believe this should be safe now.

* rust: Add Send bounds to mock transports in tests

Make sure that newly added Send bound it satisfied. The closures we use
are safe to Send across thread boundaries, but the utilities did not
preserve this bound as they themselves use type erasure.