rbthemis: Extra effort to locate Themis Core library
Just like with PyThemis, migrate RbThemis to the same pattern of loading
Themis Core library dynamically with more attention to versioning.

Detect the platform and load properly versioned libraries first,
reducing the chance for an ABI mismatch. However, keep the 'themis'
as a fallback (and a default action for Windows).

Ruby's "ffi_lib" is in some ways more smart than Python's find_library
(it supports Apple M1 by accident, for example) but in some ways it's
also dumber. Don't depend on ffi's maintainers and hardcode a list of
absolute paths to try first in attempt to make things "just work".