pythemis: Extra effort to locate Themis Core library
* pythemis: "import exception" as itself

* pythemis: "import themis" from a single place

Instead of each submodule figuring out where the native library is,
just initialize it one place and pull it from there.

* pythemis: Load canonical Themis library if possible

That is, "libthemis" ABI version 0.

First of all, this makes it possible for users to avoid installing
development package of Themis on Linux. That is, only "libthemis",
without "libthemis-dev" or "libthemis-devel".

Second, it makes sure that if Themis 0.XX and Themis 2.YY are
co-installed, then this PyThemis will load Themis 0.XX library.
By default, find_library() favors the latest version, and that
could cause ABI mismatches.

On Linux, this is how 'real' binaries load libraries via run-time
dynamic linker-loader: they remember "libthemis.so.0" soname, ask
ld to load that, and if it fails then your binary fails to start.

On macOS, applications typically remember absolute path to the library,
which would be something like this for default Homebrew install:

    /usr/local/opt/libthemis/lib/libthemis.0.dylib

However, Homebrew on Apple Silicon installs libraries in /opt/homebrew
instead of /usr/local, so look there as well.

Existing behavior of using find_library('themis') is used as a fallback
on Linux and macOS. For now, this is a genuine fallback with a warning.
PyThemis 0.15 will be a test-bed for this change, and if it works fine
then some later version might load *only* versioned libraries, assuming
libthemis.so.2 would be an actual risk.

Other systems -- like Windows -- keep using find_library(). I'm not
quite sure with the story about ABI versioning on Windows specifically.
The recommended approach there is to use absolute paths, partly because
of how easy DLL hijacking is there. Anyhow, ignore this issue for now,
and just keep using whatever we were using.

* pythemis: Handle 'library not found' errors more nicely

Turns out, if find_library() cannot find a library it simply returns
None and then LoadLibrary() does not raise an exception for that,
instead loading a "None" library, lol. That won't do.

We had people run into situations where native library could not be
found and then the interpreter reports errors like

    AttributeError: dlsym(RTLD_DEFAULT, themis_gen_sym_key): symbol not found

which confusingly *look like* some issues with library symbols.