android: Update Gradle, Android plugin, build tools, NDK
Bumping Android Gradle plugin to 7.2.2 so that it picks up new SDKs
and bumping Gradle to 7.3.3 so that it works with the new plugin,
and Android build tools to 30.0.3 so that Gradle works, and with
those new tools NDK has to be configured differently (finally,
it can be pinned in the project configuration itself, instead of
being a free variable in the build machine configuration).

This is quite a YOLO upgrade over multiple major versions, but I don't
think we have a choice. The treadmill must roll forward, because how
else we can Innovate™ to keep the number going up. Otherwise, I can't
build Android Themis anymore with today's tooling.